DOMA-3185 fixed hint display on ticket id page
