+++ apps/condo/domains/ticket/components/TicketHint/TicketHintContent.tsx
@@ -1,0 +1,13 @@
+import xss from 'xss'
+import React, { CSSProperties } from 'react'
+
+const TICKET_HINT_CONTENT_STYLES: CSSProperties = { maxHeight: '100px', overflow: 'hidden', wordBreak: 'break-word' }
+
+export const TicketHintContent = ({ ticketHint }) => (
+    <div
+        dangerouslySetInnerHTML={{
+            __html: xss(ticketHint.content),
+        }}
+        style={TICKET_HINT_CONTENT_STYLES}
+    />
+)
