+++ apps/condo/package.json
@@ -109,7 +109,8 @@
     "uuid": "^8.3.0",
     "validate.js": "^0.13.1",
     "validator": "^13.7.0",
-    "xlsx": "^0.17.0"
+    "xlsx": "^0.17.0",
+    "xss": "^1.0.13"
   },
   "devDependencies": {
     "@testing-library/react": "^11.2.5",
